+++ apps/front/src/pages/index.tsx
@@ -1,6 +1,8 @@
 import { NextPage } from 'next'
 import Head from 'next/head'
 import Link from 'next/link'
+import { useRouter } from 'next/router'
+import { useEffect } from 'react'
 import { useRooms } from '~/hooks/useRooms'
 import { useSpeechRecognition } from '~/hooks/useSpeechRecognition'
